requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
gTTS==2.2.4
idna==3.3
loguru==0.6.0
pysqlite3==0.4.7
pyTelegramBotAPI==4.7.0
python-dotenv==0.20.0
requests==2.28.1
six==1.16.0
urllib3==1.26.11
